gm内存偏移或大小使用int64



# message auto-generated for no-merge-commit merge:
!352 merge fix/offset into master

gm内存偏移或大小使用int64

Created-by: chenyanbin8
Commit-by: chenyanbin8
Merged-by: turing_project1
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->
1、masked_softmax_with_rel_pos_bias算子GM偏移量修正成int64数据类型使用
2、max_pool3d_with_argmax_v2算子无效kernel移除
## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #123-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->
ophost/opapi/kernel ut运行pass，泛化验证max_pool3d_with_argmax_v2 200用例pass
![image.png](https://raw.gitcode.com/user-images/assets/7665709/55f3fd7e-5525-43ac-8cf0-fbb7281d1378/image.png 'image.png')
## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [x] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：




See merge request: cann/ops-nn!352

+++ norm/masked_softmax_with_rel_pos_bias/op_kernel/masked_softmax_with_rel_pos_bias_B.h
@@ -86,7 +86,7 @@ class MaskedSoftmaxWithRelPosBiasB {
       BroadCastBias(attenMaskLocal);
     }
 
-    uint32_t offset = 0;
+    int64_t offset = 0;
     uint32_t stackNum = tilingData->stackNum;
     uint32_t loopNum = 0;
     if (stackNum != 0) {
@@ -95,7 +95,7 @@ class MaskedSoftmaxWithRelPosBiasB {
     // 存在loopNum，stackNum为零的情况
     uint32_t mulsNum = stackNum * wns1s2Aligned;
     SoftMaxShapeInfo softmaxShapeInfo{stackNum * wns1, s2Aligned, stackNum * wns1, s2_};
-    for (uint32_t loopCount = 0; loopCount < loopNum; loopCount++) {
+    for (int64_t loopCount = 0; loopCount < loopNum; loopCount++) {
       offset = loopCount * stackNum;
       CopyInX(offset, stackNum, tilingData->xCopyEleNum);
       if constexpr (existAtten) {
@@ -197,7 +197,7 @@ class MaskedSoftmaxWithRelPosBiasB {
     }
   }
 
-  __aicore__ inline void CopyInX(uint32_t offset, uint32_t stackNum, uint32_t copyEleNum) {
+  __aicore__ inline void CopyInX(int64_t offset, uint32_t stackNum, uint32_t copyEleNum) {
     if constexpr (isAligend) {
       LocalTensor<T> xLocal = vecInQueue.AllocTensor<T>();
       DataCopy(xLocal, xGm[offset * wns1s2Aligned], copyEleNum);
@@ -213,7 +213,7 @@ class MaskedSoftmaxWithRelPosBiasB {
     }
   }
 
-  __aicore__ inline void ComputeAttenMaskAndBias(uint32_t offset, LocalTensor<T>& attenMaskLocal, uint32_t stackNum,
+  __aicore__ inline void ComputeAttenMaskAndBias(int64_t offset, LocalTensor<T>& attenMaskLocal, uint32_t stackNum,
                                                  uint32_t mulsNum, const SoftMaxShapeInfo& softmaxShapeInfo,
                                                  const SoftMaxTiling& softmaxTilingData) {
     LocalTensor<T> xLocal = vecInQueue.DeQue<T>();
@@ -246,7 +246,7 @@ class MaskedSoftmaxWithRelPosBiasB {
     vecInQueue.FreeTensor(xLocal);
   }
 
-  __aicore__ inline void ComputeBias(uint32_t offset, LocalTensor<T>& biasLocal, uint32_t stackNum, uint32_t mulsNum,
+  __aicore__ inline void ComputeBias(int64_t offset, LocalTensor<T>& biasLocal, uint32_t stackNum, uint32_t mulsNum,
                                      const SoftMaxShapeInfo& softmaxShapeInfo, const SoftMaxTiling& softmaxTilingData) {
     LocalTensor<T> xLocal = vecInQueue.DeQue<T>();
     LocalTensor<T> yLocal = vecOutQueue.AllocTensor<T>();
@@ -270,7 +270,7 @@ class MaskedSoftmaxWithRelPosBiasB {
     vecInQueue.FreeTensor(xLocal);
   }
 
-  __aicore__ inline void CopyOut(uint32_t offset, uint32_t stackNum, uint32_t copyEleNum) {
+  __aicore__ inline void CopyOut(int64_t offset, uint32_t stackNum, uint32_t copyEleNum) {
     if constexpr (isAligend) {
       LocalTensor<T> yLocal = vecOutQueue.DeQue<T>();
       DataCopy(yGm[offset * wns1s2], yLocal, copyEleNum);
@@ -347,7 +347,7 @@ class MaskedSoftmaxWithRelPosBiasBBf16AndHalf
       BroadCastBias(attenMaskLocal, biasCastTensor);
     }
 
-    uint32_t offset = 0;
+    int64_t offset = 0;
     uint32_t stackNum = this->tilingData->stackNum;
     uint32_t loopNum = 0;
     if (stackNum != 0) {
@@ -357,7 +357,7 @@ class MaskedSoftmaxWithRelPosBiasBBf16AndHalf
     // 存在loopNum，stackNum为零的情况
     uint32_t mulsNum = stackNum * this->wns1s2Aligned;
     SoftMaxShapeInfo softmaxShapeInfo{stackNum * this->wns1, this->s2Aligned, stackNum * this->wns1, this->s2_};
-    for (uint32_t loopCount = 0; loopCount < loopNum; loopCount++) {
+    for (int64_t loopCount = 0; loopCount < loopNum; loopCount++) {
       offset = loopCount * stackNum;
       this->CopyInX(offset, stackNum, this->tilingData->xCopyEleNum);
       if constexpr (existAtten) {
@@ -449,7 +449,7 @@ class MaskedSoftmaxWithRelPosBiasBBf16AndHalf
     }
   }
 
-  __aicore__ inline void ComputeAttenMaskAndBias(uint32_t offset, LocalTensor<float>& xCastTensor,
+  __aicore__ inline void ComputeAttenMaskAndBias(int64_t offset, LocalTensor<float>& xCastTensor,
                                                  LocalTensor<float>& attenMaskLocal, LocalTensor<float>& biasCastTensor,
                                                  uint32_t stackNum, uint32_t mulsNum,
                                                  const SoftMaxShapeInfo& softmaxShapeInfo,
@@ -491,7 +491,7 @@ class MaskedSoftmaxWithRelPosBiasBBf16AndHalf
     this->vecInQueue.FreeTensor(xLocal);
   }
 
-  __aicore__ inline void ComputeBias(uint32_t offset, LocalTensor<float>& xCastTensor, LocalTensor<float>& biasLocal,
+  __aicore__ inline void ComputeBias(int64_t offset, LocalTensor<float>& xCastTensor, LocalTensor<float>& biasLocal,
                                      uint32_t stackNum, uint32_t mulsNum, const SoftMaxShapeInfo& softmaxShapeInfo,
                                      const SoftMaxTiling& softmaxTilingData) {
     LocalTensor<T> xLocal = this->vecInQueue.template DeQue<T>();
